# backend/api/v1/documents.py
from fastapi import APIRouter, UploadFile, File, Form, HTTPException, status
from typing import List, Optional
from pydantic import BaseModel

router = APIRouter()


# Response Models (TODO: Import from shared types or create Pydantic models)
class DocumentUploadResponse(BaseModel):
    documentId: str
    status: str
    estimatedProcessingTime: Optional[int] = None


class Document(BaseModel):
    id: str
    name: str
    shortName: str
    type: str
    pages: int
    status: str
    uploadedAt: str


class PaginatedDocuments(BaseModel):
    items: List[Document]
    total: int
    page: int
    pageSize: int
    hasMore: bool


@router.get("/documents")
async def list_documents(
    page: int = 1,
    pageSize: int = 20,
    type: Optional[str] = None
) -> PaginatedDocuments:
    """
    List all documents for the current user/tenant
    
    TODO: Implement:
    - Database query for documents
    - Pagination
    - Filtering by type
    - Sort by uploadedAt
    """
    # Mock response
    return PaginatedDocuments(
        items=[],
        total=0,
        page=page,
        pageSize=pageSize,
        hasMore=False
    )


@router.post("/documents", status_code=status.HTTP_202_ACCEPTED)
async def upload_document(
    file: UploadFile = File(...),
    type: str = Form(...)
) -> DocumentUploadResponse:
    """
    Upload a new PDF document for processing
    
    TODO: Implement:
    1. Validate file (PDF, size limit)
    2. Generate document ID
    3. Upload to Cloud Storage
    4. Trigger Pub/Sub for background processing
    5. Return upload response
    
    Processing pipeline (background job):
    - Extract text with PyMuPDF
    - Chunk text
    - Generate embeddings
    - Store in vector DB
    - Create Neo4j graph nodes/edges
    - Update document status to 'ready'
    """
    if not file.filename.endswith('.pdf'):
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Only PDF files are supported"
        )
    
    # TODO: Implement upload logic
    
    return DocumentUploadResponse(
        documentId="mock-doc-123",
        status="processing",
        estimatedProcessingTime=180
    )


@router.get("/documents/{documentId}")
async def get_document(documentId: str) -> Document:
    """
    Get document details by ID
    
    TODO: Implement database query
    """
    # Mock response
    raise HTTPException(
        status_code=status.HTTP_404_NOT_FOUND,
        detail="Document not found"
    )


@router.delete("/documents/{documentId}", status_code=status.HTTP_204_NO_CONTENT)
async def delete_document(documentId: str):
    """
    Delete a document
    
    TODO: Implement:
    - Delete from Cloud Storage
    - Delete from vector DB
    - Delete from Neo4j
    - Delete metadata from database
    """
    pass


@router.get("/documents/{documentId}/pages/{pageNumber}")
async def get_document_page(documentId: str, pageNumber: int):
    """
    Get rendered page image and text content
    
    TODO: Implement:
    - Fetch page image URL from Cloud Storage
    - Optionally extract text content
    - Return annotations if any
    """
    raise HTTPException(
        status_code=status.HTTP_404_NOT_FOUND,
        detail="Document or page not found"
    )