# backend/api/v1/query.py
from fastapi import APIRouter, HTTPException, status
from pydantic import BaseModel
from typing import List, Optional, Any, Dict

router = APIRouter()


# Request/Response Models
class QueryRequest(BaseModel):
    query: str
    documentIds: Optional[List[str]] = None
    filters: Optional[Dict[str, Any]] = None
    options: Optional[Dict[str, Any]] = None


class Reference(BaseModel):
    id: str
    documentId: str
    documentName: str
    page: int
    section: Optional[str] = None
    label: str
    excerpt: str
    highlightText: Optional[List[str]] = None
    confidence: Optional[float] = None


class ReasoningStep(BaseModel):
    step: int
    description: str
    action: str
    details: Optional[Dict[str, Any]] = None


class QueryResponse(BaseModel):
    queryId: str
    answer: str
    references: List[Reference]
    reasoning: Optional[List[ReasoningStep]] = None
    confidence: Optional[float] = None
    processingTime: Optional[int] = None


@router.post("/query")
async def query_documents(request: QueryRequest) -> QueryResponse:
    """
    Query documents with natural language
    
    TODO: Implement the full query pipeline:
    
    1. Generate query embedding (OpenAI embeddings API)
    2. Vector similarity search (GCP Vector Search)
       - Get top K relevant chunks
    3. Graph traversal (Neo4j Cypher queries)
       - Find related sections via cross-references
       - Traverse relationships: REFERENCES, CITES, etc.
    4. Context assembly
       - Combine vector search results + graph context
       - Rank by relevance
    5. LLM synthesis (OpenAI GPT-4)
       - System prompt: "You are an expert in construction codes..."
       - Few-shot examples for citation format
       - Include retrieved chunks as context
    6. Citation extraction
       - Parse LLM response for citations
       - Map citations back to source documents/pages
    7. Format response
       - Return answer + references + reasoning steps
    
    Example Neo4j Cypher for graph traversal:
    ```cypher
    MATCH (s:Section {id: $sectionId})
    MATCH path = (s)-[:REFERENCES|CITES*1..2]-(related:Section)
    RETURN related, path
    LIMIT 10
    ```
    
    Example LLM prompt:
    ```
    You are an expert in Australian construction codes and standards.
    Answer the user's question using ONLY the provided context.
    Always cite sources with [Doc Name, Page X, Section Y.Z].
    
    Context:
    [Retrieved chunks here]
    
    Question: {user_query}
    
    Answer:
    ```
    """
    if not request.query or len(request.query.strip()) == 0:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Query cannot be empty"
        )
    
    # Mock response for now
    return QueryResponse(
        queryId="query-123",
        answer="This is a mock response. Backend implementation pending.",
        references=[],
        reasoning=[
            ReasoningStep(
                step=1,
                description="Generated query embedding",
                action="search"
            ),
            ReasoningStep(
                step=2,
                description="Searched vector database",
                action="search"
            ),
            ReasoningStep(
                step=3,
                description="Traversed document graph",
                action="graph_traverse"
            ),
            ReasoningStep(
                step=4,
                description="Synthesized answer with LLM",
                action="synthesize"
            ),
        ],
        processingTime=1500
    )