# backend/api/v1/references.py
from fastapi import APIRouter, HTTPException, status
from pydantic import BaseModel
from typing import Optional, List, Dict, Any

router = APIRouter()


class Reference(BaseModel):
    id: str
    documentId: str
    documentName: str
    page: int
    section: Optional[str] = None
    label: str
    excerpt: str
    confidence: Optional[float] = None


class GraphNode(BaseModel):
    id: str
    type: str
    label: str
    documentId: Optional[str] = None
    page: Optional[int] = None


class GraphEdge(BaseModel):
    id: str
    source: str
    target: str
    type: str
    label: Optional[str] = None


class GraphContext(BaseModel):
    nodes: List[GraphNode]
    edges: List[GraphEdge]
    focusNodeId: Optional[str] = None


@router.get("/references/{referenceId}")
async def get_reference(referenceId: str) -> Reference:
    """
    Get reference details by ID
    
    TODO: Implement:
    - Query database for reference
    - Return full details including context
    """
    raise HTTPException(
        status_code=status.HTTP_404_NOT_FOUND,
        detail="Reference not found"
    )


@router.get("/references/{referenceId}/context")
async def get_reference_context(
    referenceId: str,
    depth: int = 2
) -> GraphContext:
    """
    Get graph context for a reference
    
    This endpoint uses Neo4j to traverse the document graph
    and find related references, citations, and cross-references.
    
    TODO: Implement Neo4j Cypher query:
    ```cypher
    MATCH (ref:Section {id: $referenceId})
    MATCH path = (ref)-[r:REFERENCES|CITES|RELATED*1..$depth]-(related)
    RETURN ref, related, r, path
    LIMIT 50
    ```
    
    Then format as nodes and edges for frontend graph visualization.
    """
    if depth < 1 or depth > 5:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Depth must be between 1 and 5"
        )
    
    # Mock response
    return GraphContext(
        nodes=[],
        edges=[],
        focusNodeId=referenceId
    )